add content_type to ContentController

+++ database/migrations/2025_07_28_191314_create_books_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
 
             $table->json('author')->nullable();
             $table->unsignedInteger('published_year')->nullable();
-            $table->integer('views')->default(0);
+            $table->unsignedBigInteger('views')->default(0);
             $table->integer('total_pages')->default(0);
             
             $table->boolean('sitemap_exclude')->nullable();
